third_party/WebKit/LayoutTests/platform/win/fast/events/platform-wheelevent-paging-x-in-scrolling-page-expected.txt

PASS event.wheelDeltaY is 0
PASS event.wheelDeltaX is window.givenScrollLeft * -3
PASS event.wheelDelta is window.givenScrollLeft * -3
PASS document.scrollingElement.scrollTop == window.expectedScrollTop && document.scrollingElement.scrollLeft == window.expectedScrollLeft became true
PASS successfullyParsed is true

TEST COMPLETE
